requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic
streamlit
pandas
pillow
openai
python-docx
lxml
mistralai
PyMuPDF
openpyxl
pyyaml
